Add CountriesDisplay component
The CountriesDisplay component takes the filtered country data from App and returns output to render depending upon its contents. If there are more than 10 results, a message is dislpayed that a different filter needs to be tried; for 2-10 results, a list of country names are displayed; for 1 result, more specific info for the country is displayed; and a message explains that nothing matches the current display for 0 results.